+++ lib/my/my.h
@@ -7,13 +7,11 @@
 ** all the functions exposed by your libmy.a
 */
 
-#include "unistd.h"
-
 #ifndef MY_H_
     #define MY_H_
 
 int my_strlen(char const *str);
-ssize_t my_putstr(char const *str);
+int my_putstr(char const *str);
 char **my_str_to_word_array(char const *str);
 int my_putnbr(int n);
 char *my_strdup(char const *src);
